# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Maratona de Resultados API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /areas:
    get:
      operationId: getAreas
      tags: [areas]
      summary: List areas
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Area"
    post:
      operationId: createArea
      tags: [areas]
      summary: Create area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AreaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
  /areas/{id}:
    patch:
      operationId: updateArea
      tags: [areas]
      summary: Update area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AreaUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
...
  /criteria:
    get:
      operationId: getCriteria
      tags: [criteria]
      summary: List criteria
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Criterion"
    post:
      operationId: createCriterion
      tags: [criteria]
      summary: Create criterion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriterionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Criterion"
  /criteria/{id}:
    patch:
      operationId: updateCriterion
      tags: [criteria]
      summary: Update criterion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriterionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Criterion"
...
  /calibrations:
    get:
      operationId: getCalibrations
      tags: [calibrations]
      summary: List calibrations
      parameters:
        - name: eventId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Calibration"
    post:
      operationId: createCalibration
      tags: [calibrations]
      summary: Create or upsert calibration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalibrationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calibration"
  /absences:
    get:
      operationId: getAbsences
      tags: [absences]
      summary: List absences
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: quarter
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Absence"
    post:
      operationId: createAbsence
      tags: [absences]
      summary: Register absence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbsenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Absence"
  /absences/{id}:
    delete:
      operationId: deleteAbsence
      tags: [absences]
      summary: Delete absence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, active]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        areaId:
          type: ["integer", "null"]
        areaName:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
...
    Area:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        active:
          type: boolean
    AreaInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    AreaUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
...
    Criterion:
      type: object
      required: [id, name, defaultWeight, active, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        responsibleAreaId:
          type: ["integer", "null"]
        responsibleAreaName:
          type: ["string", "null"]
        defaultWeight:
          type: number
        active:
          type: boolean
        displayOrder:
          type: integer
    CriterionInput:
      type: object
      required: [name, defaultWeight]
      properties:
        name:
          type: string
        description:
          type: string
        responsibleAreaId:
          type: integer
        defaultWeight:
          type: number
        displayOrder:
          type: integer
    CriterionUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        responsibleAreaId:
          type: integer
        defaultWeight:
          type: number
        active:
          type: boolean
        displayOrder:
          type: integer
...
    Calibration:
      type: object
      required: [id, eventId, employeeId, criterionId, calibratedScore, calibrationReason]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        criterionId:
          type: integer
        criterionName:
          type: string
        responsibleAreaName:
          type: ["string", "null"]
        originalAverageScore:
          type: ["number", "null"]
        calibratedScore:
          type: number
        calibrationReason:
          type: string
        calibratedByUserId:
          type: integer
        calibratedByName:
          type: string
        calibratedAt:
          type: string
    CalibrationInput:
      type: object
      required: [eventId, employeeId, criterionId, calibratedScore, calibrationReason]
      properties:
        eventId:
          type: integer
        employeeId:
          type: integer
        criterionId:
          type: integer
        calibratedScore:
          type: number
        calibrationReason:
          type: string
    Absence:
      type: object
      required: [id, employeeId, date, year, quarter, quantity]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        eventId:
          type: ["integer", "null"]
        eventName:
          type: ["string", "null"]
        date:
          type: string
        year:
          type: integer
        quarter:
          type: integer
        quantity:
          type: integer
        reason:
          type: ["string", "null"]
        registeredByUserId:
          type: integer
        createdAt:
          type: string
    AbsenceInput:
      type: object
      required: [employeeId, date, year, quarter]
      properties:
        employeeId:
          type: integer
        eventId:
          type: integer
        date:
          type: string
        year:
          type: integer
        quarter:
          type: integer
        quantity:
          type: integer
        reason:
          type: string
...